requirements.txt
opencv-python==4.5.5.64
numpy==1.21.5
pytesseract==0.3.8
Pillow==8.4.0